# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Virtual Care API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same host"
    },
    {
      "url": "https://agent.fly.dev",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Server"
    }
  ],
...
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/projects": {
      "post": {
        "operationId": "createProject",
        "parameters": [],
        "tags": ["projects"],
        "description": "Create a new project for a client.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["name", "clientId"],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "description": {
                    "type": "string"
                  },
                  "hourlyRate": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  },
                  "clientId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for 201",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "type": "object",
                      "properties": {
                        "client": {
                          "anyOf": [
                            {
                              "type": "null"
                            },
                            {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                },
                                "email": {
                                  "anyOf": [
                                    {
                                      "type": "null"
                                    },
                                    {
                                      "type": "string"
                                    }
                                  ]
                                },
                                "createdAt": {
                                  "type": "string"
                                },
                                "updatedAt": {
                                  "type": "string"
                                },
                                "userId": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "id",
                                "name",
                                "email",
                                "createdAt",
                                "updatedAt",
                                "userId"
                              ],
                              "additionalProperties": false
                            }
                          ]
                        }
                      },
                      "required": ["client"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "userId": {
                          "type": "string"
                        },
                        "description": {
                          "anyOf": [
                            {
                              "type": "null"
                            },
                            {
                              "type": "string"
                            }
                          ]
                        },
                        "hourlyRate": {
                          "anyOf": [
                            {
                              "type": "null"
                            },
                            {
                              "type": "string"
                            }
                          ]
                        },
                        "clientId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "createdAt",
                        "updatedAt",
                        "userId",
                        "description",
                        "hourlyRate",
                        "clientId"
                      ],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          },
          "404": {
            "description": "Response for 404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "enum": ["Client not found"],
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getProjects",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "clientId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "tags": ["projects"],
        "description": "Get a paginated list of projects for the authenticated user.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "type": "object",
                            "properties": {
                              "client": {
                                "anyOf": [
                                  {
                                    "type": "null"
                                  },
                                  {
                                    "type": "object",
                                    "properties": {
                                      "id": {
                                        "type": "string"
                                      },
                                      "name": {
                                        "type": "string"
                                      },
                                      "email": {
                                        "anyOf": [
                                          {
                                            "type": "null"
                                          },
                                          {
                                            "type": "string"
                                          }
                                        ]
                                      },
                                      "createdAt": {
                                        "type": "string"
                                      },
                                      "updatedAt": {
                                        "type": "string"
                                      },
                                      "userId": {
                                        "type": "string"
                                      }
                                    },
                                    "required": [
                                      "id",
                                      "name",
                                      "email",
                                      "createdAt",
                                      "updatedAt",
                                      "userId"
                                    ],
                                    "additionalProperties": false
                                  }
                                ]
                              },
                              "_count": {
                                "type": "object",
                                "properties": {
                                  "timeEntries": {
                                    "type": "number"
                                  }
                                },
                                "required": ["timeEntries"],
                                "additionalProperties": false
                              }
                            },
                            "required": ["client", "_count"],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string"
                              },
                              "updatedAt": {
                                "type": "string"
                              },
                              "userId": {
                                "type": "string"
                              },
                              "description": {
                                "anyOf": [
                                  {
                                    "type": "null"
                                  },
                                  {
                                    "type": "string"
                                  }
                                ]
                              },
                              "hourlyRate": {
                                "anyOf": [
                                  {
                                    "type": "null"
                                  },
                                  {
                                    "type": "string"
                                  }
                                ]
                              },
                              "clientId": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "createdAt",
                              "updatedAt",
                              "userId",
                              "description",
                              "hourlyRate",
                              "clientId"
                            ],
                            "additionalProperties": false
                          }
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number"
                        },
                        "pageSize": {
                          "type": "number"
                        },
                        "totalCount": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "page",
                        "pageSize",
                        "totalCount",
                        "totalPages"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["data", "pagination"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UnauthorizedErr": {
        "type": "object",
        "required": ["type", "title"],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "enum": ["Unauthorized"]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}